infra/orchestration/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harness"
version = "0.1.0"
description = "Claude Code Harness - Infrastructure orchestration for Claude development environments"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "David Garibay", email = "me@davidgaribay.dev" },
]
keywords = ["claude", "infrastructure", "ansible", "opentofu", "proxmox"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "ansible-core>=2.16.0,<2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
harness = "harness.cli:app"

[project.urls]
Homepage = "https://github.com/davidgaribay/dev-setup"
Repository = "https://github.com/davidgaribay/dev-setup"

[tool.hatch.build.targets.wheel]
packages = ["harness"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for typer)
]

[tool.ruff.lint.isort]
known-first-party = ["harness"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "typer.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=harness --cov-report=term-missing"
